ml-training/requirements.txt
# Core ML and Data Science Libraries
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
albumentations==1.3.1

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Experiment Tracking and Logging
tensorboard==2.15.1
mlflow==2.7.1
wandb==0.16.0

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2

# Development Tools
jupyter==1.0.0
ipykernel==6.26.0
black==23.9.1
flake8==6.1.0

# Optional: For advanced model architectures
timm==0.9.12  # PyTorch Image Models
efficientnet-pytorch==0.7.1

# Inference API
fastapi==0.115.0
uvicorn[standard]==0.30.6
mediapipe==0.10.11